requirements.txt
openai
langchain
dedent
unstructured
faiss-cpu
sec-api
yfinance
google-generativeai==0.3.1
google-ai-generativelanguage==0.4.0
streamlit
tiktoken
langchain_openai
langchain-google-genai
clarifai
Pillow